함수 분리 및 color.xml에서 ui.Color로 수정
