+++ app/src/main/java/com/example/compose/ui/theme/Color.kt
@@ -6,3 +6,7 @@ val Purple200 = Color(0xFFBB86FC)
 val Purple500 = Color(0xFF6200EE)
 val Purple700 = Color(0xFF3700B3)
 val Teal200 = Color(0xFF03DAC5)
+
+val PohangYellow = Color(0xFFF5CB36)
+val PohangGray = Color(0xFFEDEDED)
+val PohangAshBrown = Color(0xFF635C51)
