+++ app/src/main/res/values/colors.xml
@@ -7,7 +7,4 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="pohang_yellow">#FFF5CB36</color>
-    <color name="pohang_gray">#FFEDEDED</color>
-    <color name="pohang_ash_brown">#FF635C51</color>
 </resources>
